Add encounter/newSurvey route, minor code refactor, complete issue
