+++ packages/desktop/app/constants/encounterTabNames.js
